requirements.txt
numpy
pandas
matplotlib
requests
google-generativeai
python-dotenv